current state

+++ dhl-api/src/main/scala/com/mailytica/ai/api/freshdesk/FreshdeskId.scala
@@ -1,3 +1,3 @@
 package com.mailytica.ai.api.freshdesk
 
-case class FreshdeskId(value : Int)
+case class FreshdeskId(value : Long)
